fix: debounce 수정
